@@ -7,13 +7,12 @@
     ],
     "license": "MIT",
     "type": "library",
-    "min-stability": "stable",
     "require": {
-        "php": ">=5.6.0"
+        "php": ">=7.3"
     },
     "require-dev": {
-        "fabpot/php-cs-fixer": "^1.0",
-        "phpunit/phpunit": "^5.0"
+        "friendsofphp/php-cs-fixer": "^2.0",
+        "phpunit/phpunit": "^7.0"
     },
     "autoload": {
         "psr-4": {
