@@ -1,50 +1,79 @@
 <?php
 
 /*
- * This file is part of the Active Collab Controller project.
+ * This file is part of the ActiveCollab TemplateEngine project.
  *
  * (c) A51 doo <info@activecollab.com>. All rights reserved.
  */
 
-$header = <<<EOF
-This file is part of the Active Collab Template Engine project.
+$header = <<<'EOF'
+This file is part of the ActiveCollab TemplateEngine project.
 
 (c) A51 doo <info@activecollab.com>. All rights reserved.
 EOF;
 
-Symfony\CS\Fixer\Contrib\HeaderCommentFixer::setHeader($header);
-
-return (new Symfony\CS\Config\Config('psr2'))->fixers([
-    'header_comment',
-    'array_element_no_space_before_comma',
-    'array_element_white_space_after_comma',
-    'double_arrow_multiline_whitespaces',
-    'hash_to_slash_comment',
-    'include',
-    'join_function',
-    'multiline_array_trailing_comma',
-    'namespace_no_leading_whitespace',
-    'no_blank_lines_after_class_opening ',
-    'no_empty_lines_after_phpdocs',
-    'phpdoc_scalar',
-    'phpdoc_short_description',
-    'self_accessor',
-    'single_array_no_trailing_comma',
-    'single_blank_line_before_namespace',
-    'spaces_after_semicolon',
-    'spaces_before_semicolon',
-    'spaces_cast',
-    'standardize_not_equal',
-    'ternary_spaces',
-    'trim_array_spaces',
-    'unused_use',
-    'whitespacy_lines',
-    'ordered_use',
-    'short_array_syntax',
-    'phpdoc_params',
-    '-phpdoc_separation',
-    '-phpdoc_no_package',
-    '-print_to_echo',
-    '-concat_without_spaces',
-    '-empty_return',
-])->finder((new Symfony\CS\Finder\DefaultFinder())->in([__DIR__ . '/src', __DIR__ . '/test']));
+return (new PhpCsFixer\Config('psr2'))
+    ->setRules(
+        [
+            'header_comment' => [
+                'header' => $header,
+                'location' => 'after_open',
+            ],
+            'no_whitespace_before_comma_in_array' => true,
+            'whitespace_after_comma_in_array' => true,
+            'no_multiline_whitespace_around_double_arrow' => true,
+            'hash_to_slash_comment' => true,
+            'include' => true,
+            'no_alias_functions' => false,
+            'trailing_comma_in_multiline_array' => true,
+            'no_leading_namespace_whitespace' => true,
+            'no_blank_lines_after_class_opening' => true,
+            'no_blank_lines_after_phpdoc' => true,
+            'phpdoc_scalar' => true,
+            'phpdoc_summary' => true,
+            'self_accessor' => false,
+            'no_trailing_comma_in_singleline_array' => true,
+            'single_blank_line_before_namespace' => true,
+            'space_after_semicolon' => true,
+            'no_singleline_whitespace_before_semicolons' => true,
+            'cast_spaces' => true,
+            'standardize_not_equals' => true,
+            'ternary_operator_spaces' => true,
+            'trim_array_spaces' => true,
+            'no_unused_imports' => true,
+            'no_whitespace_in_blank_line' => true,
+            'ordered_imports' => true,
+            'array_syntax' => ['syntax' => 'short'],
+            'phpdoc_align' => true,
+            'return_type_declaration' => true,
+            'single_blank_line_at_eof' => true,
+            'single_line_after_imports' => true,
+            'single_quote' => true,
+            'phpdoc_separation' => false,
+            'phpdoc_no_package' => false,
+            'no_mixed_echo_print' => false,
+            'concat_space' => false,
+            'simplified_null_return' => false,
+            'blank_line_before_return' => true,
+            'class_attributes_separation' => [
+                'elements' => [],
+            ],
+            'linebreak_after_opening_tag' => true,
+            'native_function_casing' => true,
+            'no_closing_tag' => true,
+            'no_empty_comment' => true,
+            'no_empty_statement' => true,
+            'no_leading_import_slash' => true,
+            'lowercase_constants' => true,
+            'lowercase_cast' => true,
+            'lowercase_keywords' => true,
+        ]
+    )
+    ->setFinder(
+        (new PhpCsFixer\Finder())->in(
+            [
+                __DIR__ . '/src',
+                __DIR__ . '/test',
+            ]
+        )
+    );
