@@ -1,7 +1,7 @@
 <?php
 
 /*
- * This file is part of the Active Collab Template Engine project.
+ * This file is part of the ActiveCollab TemplateEngine project.
  *
  * (c) A51 doo <info@activecollab.com>. All rights reserved.
  */
